upload_to_bucket in hackathon_utils.py
